+++ Spolks.Mpi7/Matrices/Services/MatrixClusteringService.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using System.Numerics;
 using Matrices.Shared;
 using Matrices.Shared.Services;
 using MPI;
@@ -13,14 +12,14 @@ namespace Matrices.Mpi7.Services
         private const int MasterRank = 0;
         private const int Tag = 10;
 
-        public static Matrix2D<BigInteger> ClusteredMultiplyByAsync(this Matrix2D<BigInteger> self,
-            Matrix2D<BigInteger> multiplier)
+        public static Matrix2D<long> ClusteredMultiplyByAsync(this Matrix2D<long> self,
+            Matrix2D<long> multiplier)
         {
             Communicator communicator = Communicator.world;
             int size = communicator.Size;
             int rank = communicator.Rank;
 
-            var result = Matrix2D<BigInteger>.CreateEmpty(self.Rows, multiplier.Columns);
+            var result = Matrix2D<long>.CreateEmpty(self.Rows, multiplier.Columns);
             var rankRange = result.GetFrameIndexes(rank, size);
 
             if (rank == MasterRank)
@@ -56,14 +55,14 @@ namespace Matrices.Mpi7.Services
             return null;
         }
 
-        public static Matrix2D<BigInteger> ClusteredMultiplyBy(
-            this Matrix2D<BigInteger> self, Matrix2D<BigInteger> multiplier)
+        public static Matrix2D<long> ClusteredMultiplyBy(
+            this Matrix2D<long> self, Matrix2D<long> multiplier)
         {
             Communicator communicator = Communicator.world;
             int size = communicator.Size;
             int rank = communicator.Rank;
 
-            var result = Matrix2D<BigInteger>.CreateEmpty(self.Rows, multiplier.Columns);
+            var result = Matrix2D<long>.CreateEmpty(self.Rows, multiplier.Columns);
             var rankRange = result.GetFrameIndexes(rank, size);
             var rankResults = MultiplyFrame(rankRange.first, rankRange.last, result.Columns, self, multiplier);
