+++ Spolks.Mpi7/Mpi8/Services/MatrixGroupingClusteringService.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Numerics;
 using Matrices.Shared;
 using Matrices.Shared.Services;
 using MPI;
@@ -15,7 +14,7 @@ namespace Matrices.Mpi8.Services
     {
         private const int MasterRank = 0;
 
-        public static void MultiplyInGroups(Matrix2D<BigInteger> matrixA, Matrix2D<BigInteger> matrixB, int groups)
+        public static void MultiplyInGroups(Matrix2D<long> matrixA, Matrix2D<long> matrixB, int groups)
         {
             if (groups <= 0)
             {
@@ -29,13 +28,13 @@ namespace Matrices.Mpi8.Services
             var groupCommunicator = (Intracommunicator)world.Create(currentGroup.group);
 
             double startGroup = MPI_Wtime();
-            Matrix2D<BigInteger> matrixC = matrixA.GroupedMultiplyBy(matrixB, groupCommunicator);
+            Matrix2D<long> matrixC = matrixA.GroupedMultiplyBy(matrixB, groupCommunicator);
             double endGroup = MPI_Wtime();
 
             if (matrixC != null)
             {
                 double startNonParallel = MPI_Wtime();
-                Matrix2D<BigInteger> matrixD = matrixA.MultiplyBy(matrixB);
+                Matrix2D<long> matrixD = matrixA.MultiplyBy(matrixB);
                 double endNonParallel = MPI_Wtime();
 
                 if (world.Rank == MasterRank)
@@ -66,8 +65,8 @@ namespace Matrices.Mpi8.Services
             return groupsPerRanks;
         }
 
-        public static Matrix2D<BigInteger> GroupedMultiplyBy(this Matrix2D<BigInteger> self,
-            Matrix2D<BigInteger> multiplier, Intracommunicator communicator)
+        public static Matrix2D<long> GroupedMultiplyBy(this Matrix2D<long> self,
+            Matrix2D<long> multiplier, Intracommunicator communicator)
         {
             int size = communicator.Size;
             int rank = communicator.Rank;
@@ -76,14 +75,14 @@ namespace Matrices.Mpi8.Services
             var frameRange = MatrixDivisionService.GetFrameIndexes(resultSize, rank, size);
             int[] counts = GetCounts(resultSize, size);
 
-            BigInteger[] localResult = MatrixDivisionService.MultiplyFrame(
+            long[] localResult = MatrixDivisionService.MultiplyFrame(
                 frameRange.first, frameRange.last, multiplier.Columns,
                 self, multiplier).ToArray();
 
             if (rank == MasterRank)
             {
-                BigInteger[] globalResult = communicator.GatherFlattened(localResult, counts, MasterRank);
-                Matrix2D<BigInteger> result = Matrix2D<BigInteger>.CreateEmpty(self.Rows, multiplier.Columns);
+                long[] globalResult = communicator.GatherFlattened(localResult, counts, MasterRank);
+                Matrix2D<long> result = Matrix2D<long>.CreateEmpty(self.Rows, multiplier.Columns);
                 result.CommitFrame(0, resultSize - 1, globalResult);
 
                 return result;
